# src/main/resources/api/api.yaml
openapi: 3.0.3
info:
  title: Card Management API
  version: 1.0.0
paths:
  /api/v1/cards/register/{clientName}/{cardNumber}:
    post:
      tags:
        - Card
      summary: Cadastra um cartão para um cliente
      operationId: registerCardForClient
      parameters:
        - name: clientName
          in: path
          required: true
          schema:
            type: string
        - name: cardNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Cartões cadastrados com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'

  /api/v1/cards/import:
    post:
      tags:
        - Card
      summary: Inicia o processo de importação de cartões
      operationId: importCards
      responses:
        '202':
          description: Processamento de importação iniciado em segundo plano

  /api/v1/cards/register:
    post:
      tags:
        - Card
      summary: Cadastra um ou vários cartões
      operationId: registerCard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardRequest'
      responses:
        '200':
          description: Cartões cadastrados com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'

  /api/v1/cards/client/{clientName}:
    get:
      tags:
        - Card
      summary: Consulta os cartões de um cliente
      operationId: getCardsByClient
      parameters:
        - name: clientName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Lista de cartões do cliente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CardResponse'

components:
  schemas:
    CardRequest:
      type: object
      properties:
        clients:
          type: array
          items:
            $ref: '#/components/schemas/Client'

    CardResponse:
      type: object
      properties:
        clients:
          type: array
          items:
            $ref: '#/components/schemas/Client'

    Client:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        date:
          type: string
        cards:
          type: array
          items:
            $ref: '#/components/schemas/Card'

    Card:
      type: object
      properties:
        id:
          type: integer
        idClient:
          type: integer
        cardNumber:
          type: string
        date:
          type: string